feat: Refactor window control vuex module.
