+++ src/store/modules/ADempiere/windowControl/index.js
@@ -1,0 +1,19 @@
+
+import state from './state.js'
+import mutations from './mutations.js'
+import actions from './actions.js'
+import getters from './getters.js'
+
+/**
+ * Window Control
+ * Vuex Module
+ * @author Edwin Betancourt <EdwinBetanc0urt@outlook.com>
+ */
+const windowControl = {
+  state,
+  mutations,
+  actions,
+  getters
+}
+
+export default windowControl
